+++ src/routes/examples/usePanelRef.tsx
@@ -1,13 +1,22 @@
 import type { ComponentProps } from "react";
-import { Panel } from "react-resizable-panels";
 
-declare const props: ComponentProps<typeof Panel>;
+declare const rest: ComponentProps<typeof Panel>;
 
 // <begin>
 
-import { usePanelRef } from "react-resizable-panels";
+import { Group, Panel, usePanelRef } from "react-resizable-panels";
 
-// eslint-disable-next-line react-hooks/rules-of-hooks
-const ref = usePanelRef();
+function ExampleComponent() {
+  const ref = usePanelRef();
 
-<Panel panelRef={ref} {...props} />;
+  return (
+    <Group>
+      <Panel panelRef={ref} {...rest} />
+      {/* Other panels... */}
+    </Group>
+  );
+}
+
+// <end>
+
+export { ExampleComponent };